+++ packages/frontend/src/index.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import ReactDOM from "react-dom";
 import "./index.scss";
 import { Provider } from "react-redux";
-import { configureStore } from "./store/configureStore";
 import { MainApplication } from "./mainApplication";
+import { configureStore } from "./store";
 
 const store = configureStore();
