+++ packages/frontend/src/store/index.ts
@@ -1,3 +1,5 @@
 export * from "./application";
 export * from "./interface";
+export * from "./configureStore";
+export * from "./reducer";
 export * from "./state";
