+++ packages/frontend/src/store/interface/index.ts
@@ -1,1 +1,2 @@
+export * from "./actions";
 export { interfaceReducer } from "./reducer";
