+++ packages/frontend/src/typings/index.ts
@@ -1,0 +1,3 @@
+export * from "./data";
+export * from "./geography";
+export * from "./map";
