+++ packages/frontend/src/typings/map.ts
@@ -1,18 +1,14 @@
 import GeoJSON from "geojson";
 
 export interface IMapTopology {
+    /**
+     * Where to get fetch the topologyJSON from. In production, webpack should try and
+     * fetch this from its server.
+     */
     topologyLocation: string;
+    /**
+     * Given the json file from the topology location, this should use the topology-client package
+     * to extract the features out.
+     */
     extractFeatures: (json: any) => GeoJSON.Feature[];
-}
-
-export interface IFeature {
-    type: "Feature";
-    id: string;
-    properties: {
-        name: string;
-    };
-    geometry: {
-        type: string;
-        coorindates: Array<number[]>;
-    };
 }
